+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+  - HUB-1171
+    - Fixing migration to not die on concurrent migration with a parallel container
 
 ## [1.9.1] - 2018-06-02
 ### Added